Computer refactored in separate files
