+++ Wrappers.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "Cat.h"
 #include "VendingMachine.h"
+#include "Computer.h"
 
 struct CatWrapper
 {
@@ -14,4 +15,11 @@ struct VendingMachineWrapper
     VendingMachine* vendingMachinePtr = nullptr;
     VendingMachineWrapper(VendingMachine* ptrToVendingMachine);
     ~VendingMachineWrapper();
+};
+
+struct ComputerWrapper
+{
+    Computer* computerPtr = nullptr;
+    ComputerWrapper(Computer* ptrToComputer);
+    ~ComputerWrapper();
 };
